+++ src/CrossFade.cpp
@@ -142,7 +142,7 @@ bool CrossFader::CrossFadeMix(samplePtr buffer, sampleFormat format, sampleCount
       int clips;
       double f;
       //now, shortSeq contains the samples to mix together.
-      for (int j = 0; j < len; j++)
+      for (size_t j = 0; j < len; j++)
       {
 
          //Go through each clip
@@ -192,7 +192,7 @@ bool CrossFader::CrossFadeMix(samplePtr buffer, sampleFormat format, sampleCount
       int clips=0;
       double f;
       //Go through each sample position
-      for (int j = 0; j < len; j++) {
+      for (size_t j = 0; j < len; j++) {
 
          //go through each clip.
          for(int i= 0; i < numclips; i++)
@@ -233,7 +233,7 @@ bool CrossFader::CrossFadeMix(samplePtr buffer, sampleFormat format, sampleCount
       float f;
 
       //go through each sample position
-      for (int j = 0; j < len; j++) {
+      for (size_t j = 0; j < len; j++) {
 
 
          clips = 0;
