+++ src/effects/EffectManager.cpp
@@ -603,7 +603,7 @@ size_t EffectManager::RealtimeProcess(int group, unsigned chans, float **buffers
 
    // And populate the input with the buffers we've been given while allocating
    // NEW output buffers
-   for (int i = 0; i < chans; i++)
+   for (size_t i = 0; i < chans; i++)
    {
       ibuf[i] = buffers[i];
       obuf[i] = (float *) alloca(numSamples * sizeof(float));
@@ -620,7 +620,7 @@ size_t EffectManager::RealtimeProcess(int group, unsigned chans, float **buffers
          called++;
       }
 
-      for (int j = 0; j < chans; j++)
+      for (size_t j = 0; j < chans; j++)
       {
          float *temp;
          temp = ibuf[j];
@@ -635,7 +635,7 @@ size_t EffectManager::RealtimeProcess(int group, unsigned chans, float **buffers
    // is odd.
    if (called & 1)
    {
-      for (int i = 0; i < chans; i++)
+      for (size_t i = 0; i < chans; i++)
       {
          memcpy(buffers[i], ibuf[i], numSamples * sizeof(float));
       }
