+++ src/export/Export.cpp
@@ -1086,13 +1086,13 @@ void ExportMixerPanel::OnPaint(wxPaintEvent & WXUNUSED(event))
    angle = totAngle / ( mMixerSpec->GetNumTracks() + 1 );
 
    int max = 0, w, h;
-   for( int i = 1; i < mMixerSpec->GetNumTracks(); i++ )
+   for( size_t i = 1; i < mMixerSpec->GetNumTracks(); i++ )
       if( mTrackNames[ i ].length() > mTrackNames[ max ].length() )
          max = i;
 
    SetFont( memDC, mTrackNames[ max ], mBoxWidth, mTrackHeight );
 
-   for( int i = 0; i < mMixerSpec->GetNumTracks(); i++ )
+   for( size_t i = 0; i < mMixerSpec->GetNumTracks(); i++ )
    {
       mTrackRects[ i ].x = (int)( mBoxWidth * 2 + radius - radius *
          cos( totAngle / 2.0 - angle * ( i + 1 ) ) - mBoxWidth + 0.5 );
@@ -1120,7 +1120,7 @@ void ExportMixerPanel::OnPaint(wxPaintEvent & WXUNUSED(event))
    SetFont( memDC, wxT( "Channel: XX" ), mBoxWidth, mChannelHeight );
    memDC.GetTextExtent( wxT( "Channel: XX" ), &w, &h );
 
-   for( int i = 0; i < mMixerSpec->GetNumChannels(); i++ )
+   for( size_t i = 0; i < mMixerSpec->GetNumChannels(); i++ )
    {
       mChannelRects[ i ].x = (int)( mBoxWidth * 4 - radius  + radius *
          cos( totAngle * 0.5 - angle * ( i + 1 ) ) + 0.5 );
@@ -1141,8 +1141,8 @@ void ExportMixerPanel::OnPaint(wxPaintEvent & WXUNUSED(event))
 
    //draw links
    memDC.SetPen( wxPen( *wxBLACK, mHeight / 200 ) );
-   for( int i = 0; i < mMixerSpec->GetNumTracks(); i++ )
-      for( int j = 0; j < mMixerSpec->GetNumChannels(); j++ )
+   for( size_t i = 0; i < mMixerSpec->GetNumTracks(); i++ )
+      for( size_t j = 0; j < mMixerSpec->GetNumChannels(); j++ )
          if( mMixerSpec->mMap[ i ][ j ] )
             AColor::Line(memDC, mTrackRects[ i ].x + mBoxWidth,
                   mTrackRects[ i ].y + mTrackHeight / 2, mChannelRects[ j ].x,
@@ -1168,7 +1168,7 @@ void ExportMixerPanel::OnMouseEvent(wxMouseEvent & event)
    {
       bool reset = true;
       //check tracks
-      for( int i = 0; i < mMixerSpec->GetNumTracks(); i++ )
+      for( size_t i = 0; i < mMixerSpec->GetNumTracks(); i++ )
          if( mTrackRects[ i ].Contains( event.m_x, event.m_y ) )
          {
             reset = false;
@@ -1185,7 +1185,7 @@ void ExportMixerPanel::OnMouseEvent(wxMouseEvent & event)
          }
 
       //check channels
-      for( int i = 0; i < mMixerSpec->GetNumChannels(); i++ )
+      for( size_t i = 0; i < mMixerSpec->GetNumChannels(); i++ )
          if( mChannelRects[ i ].Contains( event.m_x, event.m_y ) )
          {
             reset = false;
@@ -1202,8 +1202,8 @@ void ExportMixerPanel::OnMouseEvent(wxMouseEvent & event)
          }
 
       //check links
-      for( int i = 0; i < mMixerSpec->GetNumTracks(); i++ )
-         for( int j = 0; j < mMixerSpec->GetNumChannels(); j++ )
+      for( size_t i = 0; i < mMixerSpec->GetNumTracks(); i++ )
+         for( size_t j = 0; j < mMixerSpec->GetNumChannels(); j++ )
             if( mMixerSpec->mMap[ i ][ j ]  && IsOnLine( wxPoint( event.m_x,
                         event.m_y ), wxPoint( mTrackRects[ i ].x + mBoxWidth,
                            mTrackRects[ i ].y + mTrackHeight / 2 ),
